poke-backend/pyproject.toml
[project]
name = "poke-backend"
version = "0.1.0"
description = "AI Agent with LangGraph, Composio, and Redis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langgraph",
    "langchain",
    "langchain-openai",
    "composio-langchain==0.8.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "asyncio-mqtt>=0.16.0",
    "celery>=5.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.scripts]
poke = "cli:main"
poke-advanced = "cli:advanced_main"
poke-server = "main:main"
